refactor(docker-configs): optimize image tag conditions and update repository paths
This commit optimizes the condition for setting Docker image tags in the bake configuration files. Instead of relying on a hard-coded version number, image tags are now determined by whether or not a VERSION environment variable is defined. If VERSION is empty, no specific version tag is assigned.

Additionally, updates were made to the Docker Compose and Firebase configuration files to change the image repositories and project default respectively, from 'utrade-taxi-run-0' to 'moov-438615'.

In the builder Dockerfile, the command to install Google Chrome and configure Bash as the default shell were reordered and dev-entrypoint.sh was moved to the end for more logical structuring of operations.

These changes allow for better flexibility and organization in our Docker configurations.