feat(Dockerfile): simplify non-root user configuration
This commit simplifies the Dockerfile instructions for setting up and giving permissions to a non-root user. We have eliminated the usage of `sudo` commands and instead temporarily switch to the `root` user directly when creating directories and installing dependencies.

Changes include:
- The `non_root_user` is now switched once to `root` for performing administrative tasks.
- The `runner` directory is created by `root`, and its ownership is granted to `non_root_user` after installation of runner dependencies.
- Removed several unnecessary empty lines, contributing to a cleaner, more readable Dockerfile.

These changes should not affect functionality but do make the Dockerfile code more straightforward and maintainable.