feat(github-actions): remove tag push step from staging workflow
This commit removes the 'Bump version and push tags' step from the GitHub Actions staging workflow. Previously, this step would increment the version number and push a new tag to the main branch. With this change, tags will no longer be automatically pushed as part of the staging process - this functionality is already done from the Github Actions production workflow.