refactor(api): change xorKey storage implementation
In the `StoreUserRecoveryXorKey` endpoint, removed the usage of Google secret manager for storing information and instead introduced a new interface `Vault` with its method `StorUserRecoveryXorKey`. This function is now responsible for storing the recovery `xorKey` for users. In addition, removed explicit dependencies in `NewService`, it now accepts the Vault instance at initialization.

In `vault.go`, added implementation of `StorUserRecoveryXorKey` which now writes the `xorKey` to a vault client, replacing previous dependency on the google secret manager.

The API consumption has changed too, 'StoreUserRecoveryXorKey' now expects a JSON body with `userId` and `xorKey`.

These changes will make the service more modular and flexible by allowing different storage mechanisms to be plugged in as needed. The error handling mechanism is also updated to provide more accurate HTTP status codes indicating the nature of the problem when something goes wrong.