chore(deps): update base Docker images and Terraform providers' versions
This commit updates the base Docker image from `debian:bookworm-stable` to `debian:bookworm`, possibly to align with the latest release or for improved stability.

Additionally, it upgrades the version of the Terraform providers 'google' and 'google-beta' from `~> 6.17.0` to `~> 6.30.0` in both the `dev` and `prod` environments. This might bring new features, bug fixes, or performance improvements.

There is also a minor formatting change in `main.tf`.

Remember to test these changes across stages to ensure nothing breaks due to introduced updates.