refactor(devops): streamline Docker image build and dev environment setup
This commit simplifies the process of building docker images and setting up the development environment. It removes redundant instructions for building individual Docker images and consolidates them under a single build instruction in `dev/docker/docker.mk`. It also adjusts the Docker compose setup, eliminating superfluous commands to start, stop, and remove services.

For the backend application, it changes the way we handle the Docker Compose. It introduces more 'make' commands for starting and deleting the environment and provides better logs on successful start. It also replaces the old backend settings with new ones, reflecting the changes made in the other files. An explicit cache write has been added to the Docker build process.
Changes were also made to Android Studio and Firebase Emulator configurations. In the case of Android Studio, this involved an adjustment to the VS Code settings persistence script. This change was necessitated due to consistency issues with the already running state.

Overall, these changes make the development workflow more streamlined and consistent.