refactor(Dockerfile): Remove clarinet installation from Dockerfile and add a README
The commit removes the previously existing section for installing Clarinet based on the target platform in the "clarinet" Dockerfile. This change suggests that future builds of the Docker image will not include the Clarinet executable.

Additionally, a new README file has been added to the "clarinet" directory, which may be intended to guide users on how to work with Clarinet outside the provided Docker environment.

Some minor changes have also been made to the tsconfig.json file, possibly for better readability.