feat(account): add account validation and sign-in using Google
This commit introduces several enhancements to the account management system. It adds a new `AccountValidate` class in the `account_validate.dart` file, which handles checking whether a Firebase user is authenticated and provides feedback about account status through SnackBars to the end user.

Also, this commit adds action buttons in `home_page_actions.dart` that route users to `WalletScreen` and `AccountPage`.

Finally, it incorporates a Google sign-in method in `sign_in_google.dart`, providing an alternative way for users to sign in or upgrade their anonymous accounts by linking them with their Google credentials. It also outlines handling for test scenarios with a Firebase emulator used in non-release mode.