feat(wallet): overhaul crypto operations and implement contract call
This commit overhauls several cryptographic operations within the wallet service. The 'math.dart' file, which contained a function for generating a random private key and XORing two keys, was deleted. The XOR operation was moved to 'wif.dart', and now operates on WIF-encoded keys instead of base64-encoded keys.

In addition, the wallet service's recovery key setup process was updated to work with WIF-encoded keys.

Finally, a new 'ContractCall' class was added in a new file 'clarity.dart'. This class represents a contract call in the Clarity smart contract language, including the address, name, and arguments of the contract being called. It also contains a simple serialization method. The wallet service now includes a method to sign and "broadcast" instances of this ContractCall class.