feat(orders): Add order service and storage
This commit introduces a new Order Service and its accompanying storage. The Order Service is set up to handle API requests related to the creation, retrieval, deletion, and listing of orders, which are stored in Firestore.

Changes introduce an 'Order' struct for capturing information related to an order such as id, name, description, sale type, weight, price, images, and date of creation. An 'OrderStorage' interface is defined to provide a consistent way of interacting with wherever the customer orders are stored - in this case, Firestore is used. A new file 'order.go' under `storage` package has also been introduced, which implements the 'OrderStorage' struct with Firestore.

The new `OrderService` is tied into the application through 'mux', a HTTP request multiplexer, and uses the FirebaseAuthMiddleware for user authentication.

Please note the new method `CreateOrder` on the `OrderService`, which handles HTTP requests for creating a new order and storing it via the OrderStorage interface. The storage was also updated to accommodate the new Order service.

This addition expands our service capabilities, providing users with the ability to manage customer orders within the application.