refactor(api): update all endpoint URLs to include '/api' prefix, modify DateTime handling in some models
In this commit, we have updated all API endpoint URLs across multiple services to use the '/api' prefix for better routing organization. The specific changes are applied throughout different files in the application including recovery_rotation_test.dart, order_service.dart, user_service.dart, vegetable_model.dart, vegetable_service.dart, vegetable_upload_provider.dart, and wallet_backend.dart.

Besides, we also made modifications to how we handle the 'createdAt' field in the Vegetable model. Instead of using a ternary condition to check if 'createdAt' is a DateTime instance before trying to convert it to a Timestamp format, we now directly convert 'createdAt' to an ISO8601 string via 'toIso8601String()' method.

In addition, for the VegetableUploadProvider, we adjusted the 'createdAt' attribute to always be in Coordinated Universal Time (UTC) by calling 'DateTime.now().toUtc()'. This ensures that irrespective of the server's timezone, the timestamp logged will be consistent and timezone-neutral.