feat(api): update models and methods related to Vegetable images
This commit updates the structure and methods used for handling Vegetable images in the API, Vegetables and Users. The 'VegetableImage' struct has been replaced by two new structs 'VegetableImageResponse' and 'VegetableImageRequest'. This allows for additional fields such as 'Path', 'ServedByCdn', 'DownloadToken' and a more detailed status enumeration. There is also a new field: 'UserCreatedAt' added to 'VegetableResponse' and 'VegetableRequest'.

This includes changes to methods in the vegetable storage & service, switching from using URL to Path. Moreover, it introduces the possibility of download tokens for vegetable images. A configuration property for enabling validated image CDN has also been introduced.

These changes provide better control and detail over the handling of vegetable image requests and responses, as well as align with the updated frontend expectations and security considerations.