feat(payment): implement Stripe payment processing
This commit introduces a new PaymentScreen and associated StripeService in the mobile application to enable Stripe payments. The PaymentScreen provides an interface for users to initiate a payment. Underlying this, the StripeService interacts with our backend server to create a checkout session with Stripe.

The new PaymentScreen communicates with the StripeService to initiate a checkout session, construct a URL, and finally launch the resulting URL using the url_launcher package. Error handling is also implemented here to handle exceptions and issues arising from failed URL launches or problems generating the checkout session.

The StripeService encapsulates the interaction with our backend server for payment-related activities. It utilizes the http package to perform a POST request to create a Stripe checkout session, providing the priceId as part of the request body. Responses are handled taking into consideration various possible HTTP status codes. It also takes care of auth headers needed for the request.

Please note that a testPriceId is currently hardcoded into the PaymentScreen to simulate a payment. This will need to be replaced dynamically according to the actual price id related to the purchase.