feat(vegetable): enhance vegetable details and sale options
This commit makes a considerable enhancement to the 'Vegetable' feature. It introduces better detail presentation for the displayed vegetables, with improved readability and information richness, such as more specific quantity and price formats. The ability to view detailed availability information about the vegetable is added.

Moreover, it expands selling options by adding an availability type feature, now allowing sellers to specify whether the vegetable is already harvested, will be harvested in the future, or will be harvested the same day. A refactoring of the Vegetable management view was also performed to optimize code reuse, thus facilitating maintenance.

It's also worth noting that the seller gallery was modified to use a new `VegetableSellerGalleryCard` component for displaying vegetable information, hence increasing code maintainability and uniformity across the app. The addition of sale-related properties and methods to the `VegetableUploadProvider` further improves flexibility when managing vegetable products.