refactor(github-workflows): simplify dev environment workflow and update various project references
This commit simplifies the Github Actions workflow for the development (`dev`) environment, reducing the process to a simple test run. This is in line with efforts to streamline our development pipeline.

The changes also include updates to several project references throughout the code base:
- In `dotenv.sh`, the GitHub Actions Runner URL has been modified.
- In `git.mk`, the commands related to handling git subtrees have been refined and some unused parts have been removed.
- The Go module path in `proxy/go.mod` has been updated to reflect the current project structure.

These updates help ensure that all components of our system are correctly linked, improving overall efficiency and maintainability.