refactor(docker): standardize image variable names and update cache configuration
This commit standardizes the naming convention of the Docker image variables across various components such as android-studio, application-tests, application-backend, application-mobile, clarinet-devnet, firebase-emulators, github-actions-runner, vault-dev. It updates the variables to have a "LOCAL_" prefix and moves the descriptor (e.g., "LATEST", "TAG") to the end of the variable for clarity.

Additionally, the commit improves the cache configuration by adding additional resources to the 'cache-from' parameter in the docker-bake.hcl files. This will allow Docker to use more local images as cache sources during the build process, potentially improving build times or reducing the amount of data that needs to be downloaded.

Finally, minor corrections were made like renaming "github-action-runner" to "github-actions-runner".