refactor(android): update local architecture, non-CI targets emulator image variables
Replaced variable usage to point to the latest android emulator image on regular developper desktop build. Also updated a directory path in the mobile images loading script. This change ensures that builds use the most recent emulator version, enhancing testing reliability with up-to-date environments. The adjustment of directory paths aims at correcting a reference for accurate data loading during local executions.
The CI build is very long because it uses groups to build images in the propper order to build all the images agains the currently builded version of them. This is accurates but takes a lot of time. In a simplier scenario, latests images of intermediary builded images to use a latest version by default. The defaults values that have been set keeps customizable by overriding of the defaults values.